+++ src/Infrastructure/Elastic/ElasticDocumentAdapter.php
@@ -19,14 +19,14 @@ final class ElasticDocumentAdapter implements DocumentAdapterInterface
         $this->client = $clientFactory->client();
     }
 
-    public function bulk(BulkOperationInterface $command)
+    public function bulk(BulkOperationInterface $command): callable|array
     {
         return $this->client->bulk([
             'body' => $command->build(),
         ]);
     }
 
-    public function update(string $index, $id, array $data)
+    public function update(string $index, $id, array $data): callable|array
     {
         return $this->client->index([
             'index' => $index,
@@ -43,18 +43,8 @@ final class ElasticDocumentAdapter implements DocumentAdapterInterface
         ]);
     }
 
-    public function flush(string $index): void
-    {
-        $matchAllQuery = [ 'query' => [ 'match_all' => (object)[] ] ];
-        $this->client->deleteByQuery([
-            'index' => $index,
-            'body' => $matchAllQuery
-        ]);
-    }
-
     public function search(SearchCommandInterface $command): Results
     {
-        $finder = new Finder($this->client, $command);
-        return $finder->find();
+        return (new Finder($this->client, $command))->find();
     }
 }